requirements.txt
dlib
rich~=14.0.0
flask~=3.1.1
numpy~=2.2.5
opencv-python~=4.11.0.86
face-recognition
mysql-connector-python~=9.3.0
pillow~=11.2.1
pyttsx3~=2.98
pypinyin~=0.54.0
Werkzeug~=3.1.3
pandas~=2.2.3
insightface~=0.7.3